Fix type mismatch

+++ src/Snapshot.c
@@ -581,7 +581,7 @@ bool dx_is_snapshot_event(const dx_snapshot_data_ptr_t snapshot_data, int event_
 }
 
 
-bool dx_is_zero_event(dx_event_id_t event_id, const dxf_event_data_t event_row, int idx) {
+bool dx_is_zero_event(dx_event_id_t event_id, const dxf_event_data_t event_row, size_t idx) {
     if (event_id != dx_eid_order)
         return false;
     return ((dxf_order_t*)event_row)[idx].size == 0;
